% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{engine_config}
\alias{engine_config}
\alias{is_engine_config}
\title{Configure the templating engine}
\usage{
engine_config(
  search_path = NULL,
  statement_open = "{\%",
  statement_close = "\%}",
  line_statement = "##",
  expression_open = "{{",
  expression_close = "}}",
  comment_open = "{#",
  comment_close = "#}",
  trim_blocks = FALSE,
  lstrip_blocks = FALSE
)

is_engine_config(x)
}
\arguments{
\item{search_path}{Path to search for template files. If \code{NULL} (the default),
searching the file system is disabled.}

\item{statement_open, statement_close}{The opening and closing delimiters
for statements (i.e. control structures).
Default: \verb{"\{\%"} and \verb{"\%\}"}.}

\item{line_statement}{The opening delimiter for an inline statement.
Default: \code{"##"}}

\item{expression_open, expression_close}{The opening and closing delimiters
for expressions (i.e. printed in rendered output).
Default: \code{"{{"} and \code{"}}"}.}

\item{comment_open, comment_close}{The opening and closing delimiters
for comments (i.e. removed from rendered output).
Default: \code{"{#"} and \code{"#}"}.}

\item{trim_blocks}{Remove first newline after a block. Default: \code{FALSE}.}

\item{lstrip_blocks}{Remove inline whitespace before a block. Default: \code{FALSE}.}

\item{x}{An object}
}
\value{
\itemize{
\item \code{engine_config()} returns a \code{"rinja_engine_config"} object.
\item \code{is_engine_config()} returns a logical scalar.
}
}
\description{
Create an object that configures how the templating engine behaves (e.g.
customize the syntax delimiters). The default values have been chosen to
match the Jinja defaults.

The equivalent Jinja class is \code{Environment}, but this term has significance
in R.
}
\examples{
engine_config()
}
