% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jinjar-package.R
\docType{package}
\name{jinjar-package}
\alias{jinjar}
\alias{jinjar-package}
\title{jinjar: Template Engine Inspired by 'Jinja'}
\description{
Template engine powered by the 'inja' C++ library. Users write a template document, using syntax inspired by the 'Jinja' Python package, and then render the final document by passing data from R. The template syntax supports features such as variables, loops, conditions and inheritance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidchall.github.io/jinjar/}
  \item \url{https://github.com/davidchall/jinjar}
  \item Report bugs at \url{https://github.com/davidchall/jinjar/issues}
}

}
\author{
\strong{Maintainer}: David Hall \email{david.hall.physics@gmail.com} (\href{https://orcid.org/0000-0002-2193-0480}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Lars Berscheid (Author of bundled inja library) [copyright holder]
  \item Niels Lohmann (Author of bundled json library) [copyright holder]
}

}
\keyword{internal}
