% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinja-package.R
\docType{package}
\name{rinja-package}
\alias{rinja}
\alias{rinja-package}
\title{rinja: Templating Engine Inspired by Jinja}
\description{
Templating engine powered by the 'inja' C++ library. Users
    write a template document, using syntax inspired by the 'jinja' Python
    package, and then render the final document by passing data from R.
    The template syntax supports features such as variables, loops and
    conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidchall.github.io/rinja}
  \item \url{https://github.com/davidchall/rinja}
  \item Report bugs at \url{https://github.com/davidchall/rinja/issues}
}

}
\author{
\strong{Maintainer}: David Hall \email{david.hall.physics@gmail.com} (\href{https://orcid.org/0000-0002-2193-0480}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Lars Berscheid (Author of bundled inja) [copyright holder]
  \item Niels Lohmann (Author of bundled json) [copyright holder]
}

}
\keyword{internal}
