% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\alias{render.character}
\alias{render.fs_path}
\title{Render template}
\usage{
render(.x, ...)

\method{render}{character}(.x, ..., .config = jinjar_config())

\method{render}{fs_path}(.x, ..., .config = jinjar_config())
}
\arguments{
\item{.x}{The template. Choices:
\itemize{
\item A template string.
\item A path to a template file (use \code{\link[fs:path]{fs::path()}}).
}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Data passed to the template.}

\item{.config}{The engine configuration (see \code{\link[=jinjar_config]{jinjar_config()}}).}
}
\value{
String containing rendered template.
}
\description{
Data is passed to a template to render the final document.
}
\examples{
# pass data as arguments
render("Hello {{ name }}!", name = "world")

# pass data as list
params <- list(name = "world")
render("Hello {{ name }}!", !!!params)

# render template file
\dontrun{
render(fs::path("template.txt"), name = "world")
}
}
