% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render template}
\usage{
render(.x, ..., .config = engine_config())
}
\arguments{
\item{.x}{The template}

\item{...}{Data passed to the template.}

\item{.config}{The engine configuration (see \code{\link[=engine_config]{engine_config()}}).}
}
\description{
Data is passed to a template to render the final document.
}
